+++ lib/ReactViews/Custom/parseCustomHtmlToReact.js
@@ -4,9 +4,7 @@ const React = require('react');
 const HtmlToReact = require('html-to-react');
 const combine = require('terriajs-cesium/Source/Core/combine');
 const defined = require('terriajs-cesium/Source/Core/defined');
-const camelCaseAttrMap = require('html-to-react/lib/camel-case-attribute-names');
-const fromPairs = require('lodash.frompairs');
-const map = require('lodash.map');
+const utils = require('html-to-react/lib/utils');
 
 const CustomComponents = require('./CustomComponents');
 
@@ -27,30 +25,6 @@ const shouldProcessEveryNodeExceptWhiteSpace = function(node) {
 };
 
 let keyIndex = 0;
-
-/**
- * @private
- */
-function createStyleJsonFromString(styleString) {
-    const toCamelCase = (str) => str.replace(/(?:^\w|[A-Z]|\b\w)/g, (ltr, idx) => idx === 0 ? ltr.toLowerCase() : ltr.toUpperCase()).replace(/\s+/g, '');
-    if (!styleString) {
-        return {};
-    }
-    const styles = styleString.split(';');
-    let singleStyle;
-    let key;
-    let value;
-    const jsonStyles = {};
-    for (let i = 0; i < styles.length; i++) {
-        singleStyle = styles[i].split(':');
-        key = toCamelCase(singleStyle[0]);
-        value = singleStyle[1];
-        if (key.length > 0 && value.length > 0) {
-            jsonStyles[key] = value;
-        }
-    }
-    return jsonStyles;
-}
 
 /**
  * @private
@@ -87,28 +61,14 @@ function getProcessingInstructions(context) {
     // Make sure any <a href> tags open in a new window
     processingInstructions.push({
         shouldProcessNode: node => node.name === 'a',
-        processNode: function(node, children) { // eslint-disable-line react/display-name
-            let elementProps = {
+        processNode: function(node, children, index) { // eslint-disable-line react/display-name
+            const elementProps = {
                 key: 'anchor-' + (keyIndex++),
                 target: '_blank',
                 rel: 'noreferrer noopener'
             };
-
-            // Process attributes
-            if (Object.keys(node.attribs).length !== 0) {
-                elementProps = combine(elementProps, fromPairs(map(node.attribs, function (value, key) {
-                    if (key === 'style') {
-                        value = createStyleJsonFromString(node.attribs.style);
-                    } else if (key === 'class') {
-                        key = 'className';
-                    } else if (camelCaseAttrMap[key]) {
-                        key = camelCaseAttrMap[key];
-                    }
-                    return [key, value || key,];
-                })));
-            }
-
-            return React.createElement('a', elementProps, node.data, children);
+          node.attribs = combine(node.attribs, elementProps);
+          return utils.createElement(node,index, node.data, children);
         }
     });
